Refactor assessmentUtils by abstracting timeline methods
It seems like a good opportunity to reduce the size of assessmentUtils.ts and instead break out all utility methods that would be used to format the assessments referral history into a timeline format.
